refactor(client): split LoadingConvertStep file
